# frontend/src/lib/openapi.json
{
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "description": "Request body for analyzing already-extracted document text.",
        "properties": {
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional source filename.",
            "title": "Filename"
          },
          "text": {
            "description": "Extracted document text to analyze.",
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
...
      "AskRequest": {
        "description": "Request body for local document question answering.",
        "properties": {
          "answer_mode": {
            "default": "auto",
            "description": "Preferred answer mode. Local LLM mode still falls back safely.",
            "enum": [
              "auto",
              "extractive",
              "local_llm"
            ],
            "title": "Answer Mode",
            "type": "string"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional source filename.",
            "title": "Filename"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional installed Ollama model override for this question.",
            "title": "Model"
          },
          "question": {
            "description": "Question to answer from the document.",
            "title": "Question",
            "type": "string"
          },
          "text": {
            "description": "Extracted document text to search.",
            "title": "Text",
            "type": "string"
          },
          "top_k": {
            "default": 5,
            "description": "Number of chunks to retrieve.",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "text",
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "Body_extract_document_documents_extract_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_extract_document_documents_extract_post",
        "type": "object"
      },
...
      "ChecklistResult": {
        "description": "Checklist and risk panel generated from document insights.",
        "properties": {
          "guidance": {
            "description": "User-facing guidance for the checklist result.",
            "title": "Guidance",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChecklistItem"
            },
            "title": "Items",
            "type": "array"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/RiskPanelItem"
            },
            "title": "Risks",
            "type": "array"
          }
        },
        "required": [
          "guidance"
        ],
        "title": "ChecklistResult",
        "type": "object"
      },
      "ClassificationResult": {
        "description": "Deterministic document classification result.",
        "properties": {
          "confidence": {
            "description": "low, medium, or high.",
            "title": "Confidence",
            "type": "string"
          },
          "document_type": {
            "description": "Best matching document category.",
            "title": "Document Type",
            "type": "string"
          },
          "evidence": {
            "description": "Document snippets that supported classification.",
            "items": {
              "type": "string"
            },
            "title": "Evidence",
            "type": "array"
          }
        },
        "required": [
          "document_type",
          "confidence"
        ],
        "title": "ClassificationResult",
        "type": "object"
      },
      "DocumentAnalysisDiagnostics": {
        "description": "Signals used to explain document-analysis confidence.",
        "properties": {
          "classification_confidence": {
            "default": "low",
            "title": "Classification Confidence",
            "type": "string"
          },
          "confidence_score": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Score",
            "type": "number"
          },
          "extracted_signal_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Extracted Signal Count",
            "type": "integer"
          },
          "grounded_field_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Grounded Field Count",
            "type": "integer"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "title": "DocumentAnalysisDiagnostics",
        "type": "object"
      },
      "DocumentInsight": {
        "description": "Structured document insights derived from extracted text.",
        "properties": {
          "action_items": {
            "items": {
              "type": "string"
            },
            "title": "Action Items",
            "type": "array"
          },
          "classification": {
            "$ref": "#/components/schemas/ClassificationResult",
            "description": "Document classification."
          },
          "confidence": {
            "description": "Overall extraction confidence.",
            "title": "Confidence",
            "type": "string"
          },
          "contact_information": {
            "items": {
              "type": "string"
            },
            "title": "Contact Information",
            "type": "array"
          },
          "diagnostics": {
            "$ref": "#/components/schemas/DocumentAnalysisDiagnostics"
          },
          "fees_or_amounts": {
            "items": {
              "$ref": "#/components/schemas/ExtractedItem"
            },
            "title": "Fees Or Amounts",
            "type": "array"
          },
          "important_dates": {
            "items": {
              "$ref": "#/components/schemas/ExtractedItem"
            },
            "title": "Important Dates",
            "type": "array"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "missing_information": {
            "items": {
              "type": "string"
            },
            "title": "Missing Information",
            "type": "array"
          },
          "official_notice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OfficialNoticeBreakdown"
              },
              {
                "type": "null"
              }
            ]
          },
          "required_documents": {
            "items": {
              "type": "string"
            },
            "title": "Required Documents",
            "type": "array"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/RiskInsight"
            },
            "title": "Risks",
            "type": "array"
          },
          "summary": {
            "description": "Short extractive summary.",
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Likely document title.",
            "title": "Title"
          }
        },
        "required": [
          "classification",
          "summary",
          "confidence"
        ],
        "title": "DocumentInsight",
        "type": "object"
      },
      "ExtractedItem": {
        "description": "A grounded extracted value with optional source text.",
        "properties": {
          "source_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Nearby source text from the document.",
            "title": "Source Snippet"
          },
          "value": {
            "description": "Extracted value.",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "ExtractedItem",
        "type": "object"
      },
...
      "ExtractionResult": {
        "description": "Result of extracting text from a document without interpretation.",
        "properties": {
          "diagnostics": {
            "$ref": "#/components/schemas/ExtractionDiagnostics"
          },
          "document_type": {
            "description": "Document extension without leading dot.",
            "title": "Document Type",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Structured error when extraction could not complete."
          },
          "filename": {
            "description": "Source filename.",
            "title": "Filename",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "description": "Local extraction metadata.",
            "title": "Metadata",
            "type": "object"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Number of pages when known.",
            "title": "Page Count"
          },
          "text": {
            "description": "Clean extracted text.",
            "title": "Text",
            "type": "string"
          },
          "warnings": {
            "description": "Non-fatal warnings.",
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "filename",
          "document_type",
          "text"
        ],
        "title": "ExtractionResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OfficialNoticeBreakdown": {
        "description": "Domain-specific breakdown for government notices and gazette notifications.",
        "properties": {
          "compliance_duties": {
            "items": {
              "type": "string"
            },
            "title": "Compliance Duties",
            "type": "array"
          },
          "confidence": {
            "default": "low",
            "description": "low, medium, or high.",
            "title": "Confidence",
            "type": "string"
          },
          "consequences": {
            "items": {
              "type": "string"
            },
            "title": "Consequences",
            "type": "array"
          },
          "effective_timing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Timing"
          },
          "gazette_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gazette Date"
          },
          "issue_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Date"
          },
          "issuing_authority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuing Authority"
          },
          "legal_basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Basis"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "notice_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice Number"
          },
          "notice_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice Type"
          },
          "payment_timing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Timing"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "table_rows": {
            "items": {
              "$ref": "#/components/schemas/OfficialNoticeTableRow"
            },
            "title": "Table Rows",
            "type": "array"
          }
        },
        "title": "OfficialNoticeBreakdown",
        "type": "object"
      },
...
      "QuestionAnswer": {
        "description": "Grounded answer produced from local document retrieval.",
        "properties": {
          "answer": {
            "description": "Answer grounded in source snippets.",
            "title": "Answer",
            "type": "string"
          },
          "confidence": {
            "description": "low, medium, or high.",
            "title": "Confidence",
            "type": "string"
          },
          "explanation": {
            "$ref": "#/components/schemas/AnswerExplanation"
          },
          "generation": {
            "$ref": "#/components/schemas/GenerationDiagnostics"
          },
          "mode": {
            "description": "Answering mode, such as extractive or local_llm.",
            "title": "Mode",
            "type": "string"
          },
          "retrieval": {
            "$ref": "#/components/schemas/RetrievalDiagnostics"
          },
          "source_snippets": {
            "items": {
              "$ref": "#/components/schemas/SourceSnippet"
            },
            "title": "Source Snippets",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "confidence",
          "mode"
        ],
        "title": "QuestionAnswer",
        "type": "object"
      },
...
      "SaveSessionRequest": {
        "description": "Request body for saving a document review session from the frontend.",
        "properties": {
          "checklist": {
            "$ref": "#/components/schemas/ChecklistResult",
            "description": "Full checklist to persist for reload."
          },
          "document_type": {
            "description": "Classified document type.",
            "title": "Document Type",
            "type": "string"
          },
          "extracted_text": {
            "description": "Extracted text used for hashing. It is retained only when PAPERVAULT_STORE_EXTRACTED_TEXT is explicitly enabled.",
            "title": "Extracted Text",
            "type": "string"
          },
          "filename": {
            "description": "Source filename.",
            "title": "Filename",
            "type": "string"
          },
          "insight": {
            "$ref": "#/components/schemas/DocumentInsight",
            "description": "Full document insight to persist for reload."
          }
        },
        "required": [
          "filename",
          "document_type",
          "extracted_text",
          "insight",
          "checklist"
        ],
        "title": "SaveSessionRequest",
        "type": "object"
      },
      "SavedSession": {
        "description": "Saved session with structured outputs and Q&A history.",
        "properties": {
          "checklist": {
            "$ref": "#/components/schemas/ChecklistResult",
            "description": "Generated checklist and risk panel."
          },
          "content_stored": {
            "default": false,
            "description": "Whether raw extracted text was explicitly retained for this session.",
            "title": "Content Stored",
            "type": "boolean"
          },
          "created_at": {
            "description": "UTC creation timestamp.",
            "title": "Created At",
            "type": "string"
          },
          "document_type": {
            "description": "Classified document type.",
            "title": "Document Type",
            "type": "string"
          },
          "extracted_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Raw extracted text when local full-content history is explicitly enabled.",
            "title": "Extracted Text"
          },
          "extracted_text_hash": {
            "description": "SHA-256 hash of extracted text.",
            "title": "Extracted Text Hash",
            "type": "string"
          },
          "filename": {
            "description": "Source filename.",
            "title": "Filename",
            "type": "string"
          },
          "insight": {
            "$ref": "#/components/schemas/DocumentInsight",
            "description": "Structured document insight."
          },
          "qa_history": {
            "items": {
              "$ref": "#/components/schemas/QuestionAnswer"
            },
            "title": "Qa History",
            "type": "array"
          },
          "session_id": {
            "description": "Local session identifier.",
            "title": "Session Id",
            "type": "string"
          },
          "text_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Always None unless explicitly configured later.",
            "title": "Text Preview"
          }
        },
        "required": [
          "session_id",
          "filename",
          "document_type",
          "extracted_text_hash",
          "created_at",
          "insight",
          "checklist"
        ],
        "title": "SavedSession",
        "type": "object"
      },
      "SessionCreatedResponse": {
        "description": "Response returned after saving a local review session.",
        "properties": {
          "id": {
            "description": "Created local session identifier.",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "SessionCreatedResponse",
        "type": "object"
      },
      "SessionMetadata": {
        "description": "Privacy-preserving metadata for a saved local document session.",
        "properties": {
          "content_stored": {
            "default": false,
            "description": "Whether raw extracted text was explicitly retained for this session.",
            "title": "Content Stored",
            "type": "boolean"
          },
          "created_at": {
            "description": "UTC creation timestamp.",
            "title": "Created At",
            "type": "string"
          },
          "document_type": {
            "description": "Classified document type.",
            "title": "Document Type",
            "type": "string"
          },
          "extracted_text_hash": {
            "description": "SHA-256 hash of extracted text.",
            "title": "Extracted Text Hash",
            "type": "string"
          },
          "filename": {
            "description": "Source filename.",
            "title": "Filename",
            "type": "string"
          },
          "session_id": {
            "description": "Local session identifier.",
            "title": "Session Id",
            "type": "string"
          },
          "text_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Always None unless explicitly configured later.",
            "title": "Text Preview"
          }
        },
        "required": [
          "session_id",
          "filename",
          "document_type",
          "extracted_text_hash",
          "created_at"
        ],
        "title": "SessionMetadata",
        "type": "object"
      },
      "SessionsClearedResponse": {
        "description": "Response returned after clearing local review history.",
        "properties": {
          "deleted": {
            "description": "Whether the local history clear completed.",
            "title": "Deleted",
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "SessionsClearedResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first document intelligence assistant.",
    "title": "Papervault",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/documents/analyze": {
      "post": {
        "description": "Analyze already-extracted text and return grounded structured insights.",
        "operationId": "analyze_extracted_document_documents_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentInsight"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Extracted Document"
      }
    },
    "/documents/ask": {
      "post": {
        "description": "Answer a question using local retrieval over provided document text.",
        "operationId": "ask_document_documents_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionAnswer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask Document"
      }
    },
    "/documents/checklist": {
      "post": {
        "description": "Generate an action checklist and risk panel from extracted document text.",
        "operationId": "create_document_checklist_documents_checklist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChecklistResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Document Checklist"
      }
    },
    "/documents/extract": {
      "post": {
        "description": "Extract text from an uploaded document without storing the original file.",
        "operationId": "extract_document_documents_extract_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_document_documents_extract_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extract Document"
      }
    },
...
    "/sessions": {
      "delete": {
        "description": "Delete every saved session and its Q&A history. Local-only, no recovery.",
        "operationId": "clear_sessions_sessions_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionsClearedResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Clear Sessions"
      },
      "get": {
        "description": "Return recent session metadata for the history page.",
        "operationId": "get_sessions_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionMetadata"
                  },
                  "title": "Response Get Sessions Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Sessions"
      },
      "post": {
        "description": "Save privacy-preserving local session history.",
        "operationId": "create_session_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session"
      }
    },
    "/sessions/{session_id}": {
      "get": {
        "description": "Return saved structured results and optional explicitly retained text.",
        "operationId": "get_session_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session"
      }
    }
  }
}